Module Typing_core_lang__Type_env

The type environment module consists of the definition of a type environment and helper functions to operate on the environment

type type_binding = Ast.Ast_types.Var_name.t * Ast.Ast_types.type_expr

type type_env = type_binding list

val check_type_equality : Ast.Ast_types.type_expr -> Ast.Ast_types.type_expr -> bool

val field_to_expr_type : Ast.Ast_types.type_field -> Ast.Ast_types.type_expr

Converts a field type to the equivalent expression type - used to check type equality

val get_var_type : Ast.Ast_types.Var_name.t -> type_env -> Ast.Ast_types.loc -> Ast.Ast_types.type_expr Core.Or_error.t

val get_class_field : Ast.Ast_types.Field_name.t -> Parsing.Parsed_ast.class_defn -> Ast.Ast_types.loc -> Ast.Ast_types.field_defn Core.Or_error.t

val get_obj_class_defn : Ast.Ast_types.Var_name.t -> type_env -> Parsing.Parsed_ast.class_defn list -> Ast.Ast_types.loc -> Parsing.Parsed_ast.class_defn Core.Or_error.t

val get_class_defn : Ast.Ast_types.Class_name.t -> Parsing.Parsed_ast.class_defn list -> Ast.Ast_types.loc -> Parsing.Parsed_ast.class_defn Core.Or_error.t

val get_function_type : Ast.Ast_types.Function_name.t -> Parsing.Parsed_ast.function_defn list -> Ast.Ast_types.loc -> (Ast.Ast_types.type_expr list * Ast.Ast_types.type_expr) Core.Or_error.t

val get_method_type : Ast.Ast_types.Function_name.t -> Parsing.Parsed_ast.class_defn -> Ast.Ast_types.loc -> (Ast.Ast_types.type_expr Core.sexp_list * Ast.Ast_types.type_expr) Core.Or_error.t

val get_function_body_expr : Ast.Ast_types.Function_name.t -> Typing_core_lang.Typed_ast.function_defn list -> Ast.Ast_types.loc -> Typing_core_lang.Typed_ast.expr Core.Or_error.t

val get_method_body_expr : Ast.Ast_types.Function_name.t -> Ast.Ast_types.type_expr -> Typing_core_lang.Typed_ast.class_defn list -> Ast.Ast_types.loc -> Typing_core_lang.Typed_ast.expr Core.Or_error.t

val get_type_capability : Ast.Ast_types.type_expr -> Typing_core_lang.Typed_ast.class_defn list -> Ast.Ast_types.loc -> Ast.Ast_types.capability Core.Or_error.t