Module Typing_core_lang.Typed_ast

type expr =

| Unit of Ast.Ast_types.loc
no need for type_expr annotation as obviously TUnit

| Integer of Ast.Ast_types.loc * int
no need for type_expr annotation as obviously TEInt

| Variable of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Var_name.t

| App of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Function_name.t * expr list

| Block of Ast.Ast_types.loc * Ast.Ast_types.type_expr * expr list
type is of the final expr in block

| Let of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Var_name.t * expr

| ObjField of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Var_name.t * Ast.Ast_types.type_expr * Ast.Ast_types.Field_name.t
First type is of the overall expr x.f, second is the type of the obj x

| ObjMethod of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Var_name.t * Ast.Ast_types.type_expr * Ast.Ast_types.Function_name.t * expr list
First type is of the overall expr x.f, second is the type of the obj x

| Assign of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Var_name.t * Ast.Ast_types.type_expr * Ast.Ast_types.Field_name.t * expr
First type is of the expr, second is the type of the obj

| Constructor of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Class_name.t * constructor_arg list

| Consume of Ast.Ast_types.loc * Ast.Ast_types.type_expr * expr
type is that of the expr being consumed

| FinishAsync of Ast.Ast_types.loc * Ast.Ast_types.type_expr * expr * expr * expr
overall type is that of the next_expr

Similar to Parsed AST, only we add an extra type_expr annotation to denote the overall type of the expression.

and constructor_arg =

| ConstructorArg of Ast.Ast_types.type_expr * Ast.Ast_types.Field_name.t * expr

type function_defn =

| TFunction of Ast.Ast_types.Function_name.t * Ast.Ast_types.type_expr * Ast.Ast_types.param list * expr

type class_defn =

| TClass of Ast.Ast_types.Class_name.t * Ast.Ast_types.cap_trait * Ast.Ast_types.field_defn list * function_defn list

Class definitions consist of the class name, the trait it is implementing (with capability associated) and the fields and methods in the class

type program =

| Prog of class_defn list * Ast.Ast_types.trait_defn list * function_defn list * expr